implement gfxtool

+++ game/gm1reader.h
@@ -29,8 +29,8 @@ namespace GM1
         GM1::Header const& Header() const;
         char const* EntryData(size_t index) const;
         size_t EntrySize(size_t index) const;
-        size_t NumEntries() const;
-        size_t NumPalettes() const;
+        int NumEntries() const;
+        int NumPalettes() const;
     };
 
 }
